backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
aiofiles==24.1.0
requests==2.32.3
Pillow==10.4.0
pyttsx3==2.98
srt==3.5.0

